Fix packaging of the namespace
Previously, the special __init__.py file that marks the namespace was
not included in the produced wheel and sdist. Since all other packages
under the namespace roseau have the file, it caused VS Code to not
recognize the installed wheel when one package has the file and another
does not. This change fixes the problem. The following have been tested:
* Wheel content: `unzip -l dist/roseau_load_flow*.whl`
* Sdist content: `tar --list -f dist/roseau_load_flow*.tar.gz`
* Uninstall: `pip install` and `pip uninstall` of the package work as
  expected -- uninstall does not affect other subpackages in the same
  namespace