Implement the generic load
The interface is ready for a simple way to define a generic load
with either a constant power, a constant current, a constant
impedance, or a combination of them. Currently only constant power
or constant impedance loads are impelmented in the backend.

The equation of the load is in the `Load` constructor doc in the API.
In a future PR, the equation will be added to the tutorial as well.