Improve catalogue printing on dark background and narrow screens (#135)
- Use more generic colors that work on both dark and light backgrounds.
The colors are taken from the default palette in seaborn.
- Change the overflow mechanism to fold instead of cropping the entries
in the table. This works better on narrower screens.
- Color by row instead of by column. This looks better in the case where
an entry spans multiple rows.

I also fix the console width to 210 columns in tests to get
deterministic results because depending on the width, some rows might
fold to next line. Now there is one line per catalogue entry in the
table.