Add voltage unbalance calculation (#142)
Closes #136

Add method `res_voltage_unbalance` to 3-phase buses that calculates the
voltage unbalance factor. Note that this is a method, not a property, so
that we can support more standards in the future by passing arguments to
this method.

Voltage unbalance does not make sense for non 3-phase buses thus we
raise an error with a helpful message.

I also fixed the symmetrical to phasor converters to preserve the shape
of the input array.