Fix typing of numpy arrays (#144)
This PR fixes all annotations of numpy array and makes it explicit what
type of array a function expects/returns.

We were using invalid numpy array annotations in some places and didn't
precise the type of the array in other places. To properly type numpy
array one needs to write this notation `np.ndarray[Any,
np.dtype[np.complex_]]` where the first type is the shape of the array
and the second is its type. This is not practical and is why numpy
introduced `numpy.typing.NDArray` which can be simply used like
`NDArray[np.complex_]`. Because we use this type a lot, I added a type
alias.