Improve the typing of pint wrapped method
This PR improves the user facing types of parameters of methods
wrapped with pint wrappers. Inside the methods, the types of parameters
that accept either a quantity or a unitless value become incorrect but users
of these methods now get the correct types. This means we accept a little
bit of wrong types in our internal code for the benefit of our users. Until
the Python type system have a "Map" type support, this is our best option.