Optimize results (#192)
This PR includes some optimizations that together added up to 3x
performance improvements in a timeseries study I am doing especially on
line results. The main things done are:
1. Caching whenever possible for things that we know are immutable like
phases, z_line and its inverse matrices, voltage_phases, etc.
2. Avoiding the use of pint `.m_as` as much as possible. When using pint
quantities, accessing the value using `.m` is much faster in tight loops
when we know that the unit we want is the default one. Also avoiding the
whole pint conversion interface (using `ureg_wraps` in our case)
contributed the most to the performance improvements as it avoids the
double conversion to a quantity then to a value.
3. Inlining results access in the `res_` properties on the electrical
network to avoid calling `_res_currents_getter` several times for
example.
4. A few other micro-optimizations