Explicitly prevent the instantiation of abstract classes (#202)
Closes #188 

Inheriting from `ABC` (or using `ABCMeta` as a metaclass) prevents the
instantiation of classes that have abstract methods but not classes
without abstract methods. This PR improves the error in case someone
tries to instantiate `Element`, `AbstractLoad`, or `AbstractBranch`. I
used the same error type and message used by `ABC`.