Allow floating neutrals for loads and sources (#227)
This was always allowed internally with a private flag but is now
exposed to the users.
The only behavioral change is how we get the results of the potentials
of the load. Previously we used the potentials of the bus connected to
the load. This cannot be always used now as a floating neutral means
that the neutral does not exist for the bus and its potential must be
obtained from the load or source element directly. The change remains
internal and has no effect on currently valid code or JSON files.