Improve asssigning of potential refs in DGS networks (#230)
This PR implements a better algorithm for auto assigning potential
references to DGS networks.
It replaces the heuristic used earlier that worked for simple networks
like MV/LV network with one transformer, LV network with neutral and MV
network without neutral to work with any network configuration.

The new algorithm is based on the method `_check_ref` we use to check
the correctness of potential references but it adds potential references
instead of raising an error. The algorithm is currently DGS specific it
makes some assumptions that may not be true in the general case.

We might want to consider adding a similar functionality for the general
case later though. This could be useful for users who simply don't care
about this low level functionality or for use in a graphical user
interface.