Add `connect_neutral` parameters to loads and sources
This gives the users the ability to create loads with connected neutral
and loads with floating neutral on the same bus without needing to create
a dedicated "abc" bus just to have a floating neutral connection.

The default behavior of automatically considering the neutral as floating
when the bus does not have a neutral is conserved.