Add DGS import/export to load_flow_single (#342)
Towards #334

This PR adds import/export to DGS format to the single-phase library.
Export files have been successfully loaded into PwF and the load flow
succeeded. A test case is written to ensure roundtrip conversion to dgs.
More unit test could be added later to cover edge cases.

There are a couple of changes that need to be made in the multi-phase
library but they will have to wait for now.