[single] Return scalar results as Python objects (#367)
This is a purely cosmetic change to improve the readability of the
results so that when we introspect the results in an interactive console
like a notebook cell we see something like `(-43.17581696+81.68999335j)`
instead of `np.complex128(-43.17581696+81.68999335j)` which is even more
annoying when seen in the dict of a network. Performance is not affected
by this change.