Add results of ground potential to shunt line (#369)
This uses the newly added C++ branch methods to access the potential of
the ground port of a line with shunt componenets. This makes the line
results accessible independent of the Python ground element. No breaking
changes are made to the line results interface but the network JSON
format needed to be updated to store the new results.

This work will be essential later when we simplify ground handling in
RLF.