Improve conversion of asymmetric multi-phase lines to single-phase lines (#384)
Instead of using the average of the matrix elements, we use the diagonal
part of the positive sequence elements in the symmetrical component
matrix. In general, this is a more principled approach than arbitrarily
taking averages, even though the difference is not enormous.
We could add a `rlf.LineParameters.to_sym` method later that applies the
same logic here after we do some more real world testing.

See also:
https://www.roseautechnologies.com/theorie-des-composantes-symetriques-lignes-equilibrees/