Stop reformatting arrays in json output (#387)
It is not needed anymore as we use biome to format the json files. This
also improves stability of json serialization because arbitrary strings
can be used as element IDs which could fool the regular expression.

Also change `dumps` to `dump` for slightly better memory efficiency as
it does the writing in chunks.