Fix disconnecting loads and sources (#392)
The following changes are made:
- Disconnect a ground connection of a load or source when the load or
source is disconnected from the network
- Raise a clear error message when accessing any result of a
disconnected element
- Add `is_disconnected` property to loads, sources and ground
connections to check if the element is disconnected
- In the future, accessing the bus of a disconnected load or source will
return the original bus instead of `None`. Use `is_disconnected` to
check if the load or source is disconnected.

About changing the bus to no longer be `None` for disconnected elements,
I find myself needing to access the bus in many cases mainly to connect
another element to it. The new behavior works well as it guarantees a
clear error message when the user tries to access the results of a
disconnected element while keeping access to its bus.