Make the graph data of the network json serializable (#403)
I am getting a `TypeError: Object of type Point is not JSON serializable`
when trying to plot the graph of the network with pyviz.
This is because we store the geometries as shapely objects in the graph.
This PR changes the geometries to geojson-like format so that they are
JSON serializable for better compatibility with the networkx ecosystem.
I also noticed that the buses were missing some data so I added them.